ref: externalized OpenCloseButton from MobileSideBar
